MusicHub/requirements.txt
aiofiles==24.1.0
aiogram==3.17.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
asyncio==3.4.3
attrs==25.1.0
backoff==1.11.1
beautifulsoup4==4.13.3
certifi==2025.1.31
charset-normalizer==3.4.1
colorama==0.4.6
distro==1.9.0
Django==4.2.19
django-bootstrap-v5==1.0.11
filelock==3.17.0
frozenlist==1.5.0
fsspec==2025.2.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.30.1
idna==3.10
Jinja2==3.1.5
jiter==0.9.0
Levenshtein==0.27.1
logging==0.4.9.6
magic-filter==1.0.12
MarkupSafe==3.0.2
matlab==0.1
mpmath==1.3.0
multidict==6.1.0
networkx==3.4.2
numpy==2.2.2
openai==1.68.2
packaging==24.2
pillow==11.1.0
propcache==0.2.1
pydantic==2.10.6
pydantic_core==2.27.2
pygame==2.6.1
PyMySQL==1.1.1
python-Levenshtein==0.27.1
PyYAML==6.0.2
RapidFuzz==3.12.2
redis==5.2.1
regex==2024.11.6
requests==2.32.3
safetensors==0.5.2
setuptools==75.8.0
sniffio==1.3.1
soupsieve==2.6
spotify==0.10.2
spotipy==2.25.1
SQLAlchemy==2.0.37
sqlparse==0.5.3
sympy==1.13.1
tokenizers==0.21.0
torch==2.6.0
torchaudio==2.6.0
torchvision==0.21.0
tqdm==4.67.1
transformers==4.51.0
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
Wikipedia-API==0.8.1
yarl==1.18.3
